+++ CHANGELOG.md
@@ -30,6 +30,7 @@ All Notable changes to `cron` will be documented in this file
 - `RangeError` exception is removed as it is no longer needed.
 - **[BC Break]** `ExpressionParser` its capabilities are now bundle inside the `Expression` class.
 - **[BC Break]** `CronFieldValidator` capabilities are removed and replaced by `CronField` value objects.
+- **[BC Break]** `Expression::__toString`, The `Expression` class no longer implements the `Stringable` interface.
 
 ## [0.4.0] - 2022-01-03